Refactor integration tests to separate files
