+++ package.json
@@ -70,7 +70,8 @@
     "mocha": "~1.18",
     "nock": "~0.28.2",
     "proxyquire": "~0.6.0",
-    "coveralls": "~2.10.0"
+    "coveralls": "~2.10.0",
+    "node-uuid": "~1.4.1"
   },
   "scripts": {
     "test": "grunt test"
