+++ test/commands/index.js
@@ -1,0 +1,5 @@
+describe('integration tests', function () {
+    require('./init');
+    require('./install');
+    require('./uninstall');
+});
